Done implementing the computer side heap sort algorithm
The algorithm works and it's being displayed properly. I removed some dependencies in on of the useEffects that was causing problems with the rendering of the lines. I added the swapping animations and also the nodes change color when the value has been swapped after the first stage of the heap sort algorithm.