Update MergeSort.jsx
Added difficultyTimeInterval state to be used by the algorithm. I have part of the merge sort algorithm working. I still need to make some adjustments and make it work with the UI. I added className valueToSort and valueToSort${index} because I'll be using a grid to display the UI corerclty.